+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/curse/CurseOfUnstableMotion.java
@@ -43,8 +43,8 @@ public class CurseOfUnstableMotion extends Enchantment {
         if (event.getEntity().level().isClientSide)
             return;
 
-        if (EnchantmentHelper.getEnchantmentLevel(NewEnchantmentsFeature.CURSE_OF_UNSTABLE_MOTION.get(), event.getEntity()) <= 0
-                && event.getEntity().getAttribute(Attributes.MOVEMENT_SPEED).getModifier(MODIFIER_UUID) != null) {
+        if (EnchantmentHelper.getEnchantmentLevel(NewEnchantmentsFeature.CURSE_OF_UNSTABLE_MOTION.get(), event.getEntity()) <= 0) {
+            if (event.getEntity().getAttribute(Attributes.MOVEMENT_SPEED).getModifier(MODIFIER_UUID) != null)
             event.getEntity().getAttribute(Attributes.MOVEMENT_SPEED).removeModifier(MODIFIER_UUID);
         }
         else {